+++ tools/themes/tool.js
@@ -10,7 +10,7 @@
         type: 'multiple',
         possibleValues: ['performance', 'justExplorer', 'mobile', 'slimArrows', 'slimmerArrows', 'flairX', 'lessIcons', 'nonStickyHeader', 'toggleStudyChat',
                          'pieceDrag','noPractice', 'gameMoveList', 'fatGauge', 'fatMove', 'gridBoard','noResultPopup','adamisko','fixThirdParties','timeControls',
-                         'mobileHideStartEndButtons'],
+                         'mobileHideStartEndButtons', 'mobileNoFish'],
         defaultValue: 'fixThirdParties',
         advanced: true
       }
@@ -40,7 +40,8 @@
         'themes.adamisko': 'Vintage Adamisko',
         'themes.fixThirdParties': 'Fix third parties',
         'themes.timeControls': 'Hover time controls',
-        'themes.mobileHideStartEndButtons': 'Mobile hide start/end buttons'
+        'themes.mobileHideStartEndButtons': 'Mobile hide start/end buttons',
+        'themes.mobileNoFish': 'No fish button on mobile'
       },
       'ro-RO': {
         'options.appearance': 'Aspect',
@@ -64,7 +65,8 @@
         'themes.noResultPopup': 'F\u0103r\u0103 popup cu rezultat joc',
         'themes.fixThirdParties': 'Repar\u0103 ter\u0163e par\u0163i',
         'themes.timeControls': 'Controale timp la hover',
-        'themes.mobileHideStartEndButtons': 'Ascunde butoanele start/final pe mobil'
+        'themes.mobileHideStartEndButtons': 'Ascunde butoanele start/final pe mobil',
+        'themes.mobileNoFish': 'F\u0103r\u0103 buton cu pe\u015fte pe mobil'
       }
     }